server/requirements.txt
en-core-web-sm>=3.7.1
fastapi>=0.110.0
motor>=3.3.2
python-dotenv>=1.0.1
redis>=5.0.2
scikit-learn==1.4.1.post1
spacy>=3.7.4
uvicorn>=0.27.1
